+++ frontend/lib/ui/chat/widget/chat_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../chatbot_view_model.dart';
+import '../view_model/chatbot_view_model.dart';
 
 class ChatWidget extends StatelessWidget {
   const ChatWidget({super.key});
