+++ frontend/lib/ui/chat/widget/chatbot_send_message_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../chatbot_view_model.dart';
+import '../view_model/chatbot_view_model.dart';
 
 class ChatbotSendMessageWidget extends StatefulWidget {
   const ChatbotSendMessageWidget({super.key});
