+++ frontend/test/ui/chat/widget/chat_screen_test.dart
@@ -3,12 +3,12 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
 
 // Sostituisci i percorsi in base al tuo progetto
-import 'package:mvp_app_protegge_e_trasforma/presentation/chat/chat_screen.dart';
-import 'package:mvp_app_protegge_e_trasforma/presentation/chat/chatbot_view_model.dart';
-import 'package:mvp_app_protegge_e_trasforma/presentation/chat/widgets/chat_widget.dart';
-import 'package:mvp_app_protegge_e_trasforma/presentation/chat/widgets/chatbot_send_message_widget.dart';
+import 'package:mvp_app_protegge_e_trasforma/ui/chat/widget/chat_screen.dart';
+import 'package:mvp_app_protegge_e_trasforma/ui/chat/view_model/chatbot_view_model.dart';
+import 'package:mvp_app_protegge_e_trasforma/ui/chat/widget/chat_widget.dart';
+import 'package:mvp_app_protegge_e_trasforma/ui/chat/widget/chatbot_send_message_widget.dart';
 
-import '../../../testing_utilities/mocks/mock_chatbot_repository.dart';
+import '../../../../testing/mocks/mock_chatbot_repository.dart';
 
 void main() {
   group('ChatScreen (Integration UI Test)', () {
