+++ frontend/test/ui/chat/widget/chatbot_create_chat_widget_test.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
-import 'package:mvp_app_protegge_e_trasforma/presentation/chat/widgets/chatbot_create_chat_widget.dart'; // Adatta il nome del file se diverso
-import 'package:mvp_app_protegge_e_trasforma/presentation/chat/chatbot_view_model.dart';
-import '../../../../testing_utilities/mocks/mock_chatbot_repository.dart';
+import 'package:mvp_app_protegge_e_trasforma/ui/chat/widget/chatbot_create_chat_widget.dart'; // Adatta il nome del file se diverso
+import 'package:mvp_app_protegge_e_trasforma/ui/chat/view_model/chatbot_view_model.dart';
+import '../../../../testing/mocks/mock_chatbot_repository.dart';
 
 void main() {
   group('ChatbotCreateChatWidget Widget Test', () {
