+++ frontend/testing/mocks/mock_chatbot_service.dart
